Flytter all MDC logging til en fil og legger på correlationId på oppslag.
Serverkall -> hente correlation-id header eller opprett ny og legg i MDC.
Klientkall -> hent correlation-id fra MDC eller opprett ny og legg på request header.
Kafka konsumering -> hent correlation id fra payload eller opprett ny og legg i MDC.
