+++ src/main/kotlin/no/nav/sifinnsynapi/oppslag/OppslagsKlientKonfig.kt
@@ -2,7 +2,7 @@ package no.nav.sifinnsynapi.oppslag
 
 import no.nav.security.token.support.client.core.oauth2.OAuth2AccessTokenService
 import no.nav.security.token.support.client.spring.ClientConfigurationProperties
-import no.nav.sifinnsynapi.http.MDCValuesPropagatingClienHttpRequesInterceptor
+import no.nav.sifinnsynapi.filter.MDCValuesPropagatingClienHttpRequesInterceptor
 import no.nav.sifinnsynapi.util.Constants.X_CORRELATION_ID
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -46,7 +46,6 @@ class OppslagsKlientKonfig(
         return builder
                 .setConnectTimeout(Duration.ofSeconds(20))
                 .setReadTimeout(Duration.ofSeconds(20))
-                .defaultHeader(X_CORRELATION_ID, UUID.randomUUID().toString())
                 .defaultHeader(CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
                 .rootUri(oppslagsUrl)
                 .defaultMessageConverters()
