+++ src/main/kotlin/no/nav/sifinnsynapi/saf/SafService.kt
@@ -5,6 +5,8 @@ import com.expediagroup.graphql.client.types.GraphQLClientResponse
 import com.fasterxml.jackson.databind.ObjectMapper
 import no.nav.sifinnsynapi.saf.generated.HentJournalpostinfo
 import no.nav.sifinnsynapi.saf.generated.hentjournalpostinfo.Journalpost
+import no.nav.sifinnsynapi.util.Constants.CORRELATION_ID
+import no.nav.sifinnsynapi.util.MDCUtil
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 
@@ -19,7 +21,9 @@ class SafService(
 
     suspend fun hentJournalpostinfo(journalpostId: String): Journalpost {
         val response: GraphQLClientResponse<HentJournalpostinfo.Result> =
-            safClient.execute(HentJournalpostinfo(HentJournalpostinfo.Variables(journalpostId)))
+            safClient.execute(HentJournalpostinfo(HentJournalpostinfo.Variables(journalpostId))) {
+                header(CORRELATION_ID, MDCUtil.callIdOrNew())
+            }
 
         return when {
             response.data!!.journalpost != null -> {
